Refactored TimerSpec. Fixed scalajs implicit which had been set to null.
