+++ odelay-core/js/src/main/scala/platform.scala
@@ -1,5 +1,5 @@
 package odelay
 
 private[odelay] object platform {
-  implicit val defaultTimer: Timer = null
+  implicit val defaultTimer: Timer = odelay.js.JsTimer.newTimer 
 }
